requirements.txt
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
biopython==1.80
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.0
Flask==2.2.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
outcome==1.2.0
pycodestyle==2.10.0
pycparser==2.21
PySocks==1.7.1
requests==2.28.1
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
Werkzeug==2.2.2
wsproto==1.2.0
